% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissuesAnalysis.R
\name{runAndSave_tissueAnalysis_components}
\alias{runAndSave_tissueAnalysis_components}
\title{Second wrapper to compute and (optionally) save a list of logPhv = CpG x dataset matrix of
log(Phv|Dk), numerator = CpG x dataset matrix of numerator, denominator = CpG x
dataset matrix of denominator, dataset_names = dataset names}
\usage{
runAndSave_tissueAnalysis_components(
  analysis,
  cpg_names_vec,
  resultDir,
  NCORES,
  p1,
  overwrite = FALSE,
  batch_size = 1000L,
  dataDir,
  skipsave = FALSE,
  Nds = 3L,
  subsetMetadata = FALSE
)
}
\arguments{
\item{analysis}{Character string. Name of the analysis.
If it contains \code{"MariasarraysREDUCED"}, a special directory structure is expected.}

\item{cpg_names_vec}{Character vector of CpG identifiers to process.}

\item{resultDir}{Character string; output directory for result files.}

\item{NCORES}{Integer; number of CPU cores to use in parallel processing.}

\item{p1}{Numeric scalars: True positive rate.}

\item{overwrite}{Logical; if \code{TRUE}, overwrite existing result file (default FALSE).}

\item{batch_size}{Integer; number of CpGs to process per batch (default 10000).}

\item{dataDir}{Character string. Path to the directory containing input data files.
Should contain \code{sample_metadata.tsv}, \code{all_medsd_lambda.tsv}, and an HDF5 matrix file.}

\item{skipsave}{Logical; if \code{TRUE}, skip saving results to disk.}

\item{Nds}{Integer; minimum number of datasets required per CpG (default 3).}

\item{subsetMetadata}{Logical or data frame. If \code{FALSE} (default), the full metadata
is used. Otherwise, a subset of the metadata can be provided to restrict
the analysis to specific samples or datasets.}
}
\value{
Invisibly returns the result list \verb{(CpG x dataset)} of \code{log(Phv|Dk)} ,
numerator and denominator values.
Also saves an \code{.RData} file in \code{resultDir} unless \code{skipsave = TRUE}.
}
\description{
Second wrapper to compute and (optionally) save a list of logPhv = CpG x dataset matrix of
log(Phv|Dk), numerator = CpG x dataset matrix of numerator, denominator = CpG x
dataset matrix of denominator, dataset_names = dataset names
}
